4.21 saakka

+++ vk4/blogilista/controllers/blogs.js
@@ -1,10 +1,19 @@
 const blogsRouter = require('express').Router();
+const jwt = require('jsonwebtoken');
 const Blog = require('../models/blog');
 const User = require('../models/user');
 
 function badRequest(res, errorText) {
     return res.status(400).json({ error: errorText });
 }
+
+const getTokenFrom = (request) => {
+    const authorization = request.get('authorization');
+    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+        return authorization.substring(7);
+    }
+    return null;
+};
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -22,8 +31,18 @@ blogsRouter.get('/:id', async ({ params: { id } }, response) => {
         : response.status(404).json({ error: 'no such id' });
 });
 
-blogsRouter.post('/', async ({ body }, response) => {
+blogsRouter.post('/', async (request, response) => {
+    const { body } = request;
+
     try {
+        const { token } = request;
+        const decodedToken = jwt.verify(token, process.env.SECRET);
+
+        if (!token || !decodedToken.id) {
+            return response.status(401)
+                .json({ error: 'token missing or invalid' });
+        }
+
         if (!body.title || !body.url || !body.userId) {
             return badRequest(response, 'mandatory parameter missing');
         }
@@ -41,14 +60,24 @@ blogsRouter.post('/', async ({ body }, response) => {
 
         return response.status(201).json(savedBlog);
     } catch (err) {
+        if (err.name === 'JsonWebTokenError') {
+            return response.status(401).json({ error: err.message });
+        }
+
         console.log(err);
         return response.status(500).json({ error: 'something went wront' });
     }
 });
 
-blogsRouter.delete('/:id', async ({ params: { id } }, response) => {
+blogsRouter.delete('/:id', async ({ params: { id }, token }, response) => {
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+    const blog = await Blog.findById(id);
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+        return response.status(401).json({ error: 'soosoo' });
+    }
+
     await Blog.findByIdAndRemove(id);
-    response.status(204).end();
+    return response.status(204).end();
 });
 
 blogsRouter.put('/:id', async ({ body, params: { id } }, response) => {
